Introduced Vector2D and also deleted redundant cas library

+++ src/math/Prime.java
@@ -1,10 +1,10 @@
 package math;
 
 public class Prime {
-	public static boolean isPrime(int num) {
+	public static boolean isPrime(long num) {
 		if (num < 2)
 			return false;
-		for (int i = 2; i <= java.lang.Math.sqrt(num); i++) {
+		for (long i = 2; i <= java.lang.Math.sqrt(num); i++) {
 			if (num % i == 0)
 				return false;
 		}
